340-ce-nextgen/src/test/java/io/harness/ccm/remote/resources/HealthStatusTest.java
package io.harness.ccm.remote.resources;

import static io.harness.annotations.dev.HarnessTeam.CE;

import static org.assertj.core.api.Assertions.assertThat;

import io.harness.CategoryTest;
import io.harness.annotations.dev.OwnedBy;
import io.harness.category.element.UnitTests;
import io.harness.ng.core.Status;
import io.harness.ng.core.dto.ResponseDTO;
import io.harness.rule.Owner;
import io.harness.rule.OwnerRule;

import java.io.IOException;
import org.junit.Before;
import org.junit.Test;
import org.junit.experimental.categories.Category;
import org.mockito.InjectMocks;
import org.mockito.MockitoAnnotations;

@OwnedBy(CE)
public class HealthStatusTest extends CategoryTest {
  @InjectMocks HealthStatus healthStatus;

  @Before
  public void setUp() throws Exception {
    MockitoAnnotations.initMocks(this);
  }

  @Test
  @Owner(developers = OwnerRule.UTSAV)
  @Category(UnitTests.class)
  public void testBaseGetMethod() throws IOException {
    ResponseDTO<Boolean> responseDTO = healthStatus.status();
    assertThat(responseDTO.getStatus()).isEqualTo(Status.SUCCESS);
    assertThat(responseDTO.getData()).isTrue();
  }
}